merge pending inserts into row groups
Pending inserts are inserted into the primary index as row-oriented records. When the number of pending inserts in a node of the primary index exceeds a threshold, the pending inserts are merged into the row group of the node (if there is one). Merging the pending inserts creates 1 or more new row groups from the records in the existing row group of the node and the pending inserts in the node. If the node does not have a row group, the new row group(s) are created from the pending inserts only.