use a buffered writer when writing stripes
Switch the stripe writers to use the generic `std.io.Writer` interface instead of the blob "interface". Have the segment wrap each stripe writer in a buffered writer so all writes to the blob are buffered. This significantly improves writer performance. Modify the interface for initializing a segment writer to remove the need for copying the large buffers.