coerce values on insert into primary index
The primary index uses columns of type ANY to store values for pending inserts. When the data being inserted comes from a csv, all values are inserted as strings. This means that these values lived in the primary index as strings. When the values are inserted into row groups (before and after this commit), the `sqlite3_value_*` function is called based on the column type, which also coerced values where necessary. This created a mismatch where values could be in the primary index as strings and in the row groups as the proper type.

To ensure consistency between values in the primary index and values in the row groups, and to report type errors to the user immediately on insert, coerce values on insert into the primary index.